requirements.txt
fabric
python-vagrant
micromodels
git+git://github.com/zbyte64/micromodels-collections.git#egg=micromodels-collections
